server/requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
networkx==3.2.1
openai==1.81.0
pathlib==1.0.1
pathspec==0.12.1
protobuf==4.25.7
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sievedata==1.4.12
six==1.17.0
sniffio==1.3.1
sseclient==0.0.27
tqdm==4.67.1
typeguard==4.4.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
